URL to HTML Hyperlink
If you have a text list of URLs, you can automatically convert those URLs to HTML hyperlinks using this online tool.

If you ever receive a list of multiple URLs that you want to quickly convert into HTML links to display on a web page then this the tool is for you to produce clickable links.

It will convert the list of web addresses to hyperlinks if they are formatted in one of three ways. Either the URLS are separated by commas or spaces or each of the URLs is on its own separate line

It's pretty versatile in taking lots of differently formatted information source for the URLs and changing into html code you can actually use to publish.

You can also have this tool to generate code that will enclose the hyperlinks in a HTML paragraph tag or put them in an ordered list. These options will make it easier to copy and paste the URL list into a webpage if that's what you need to do.